backend/src/requirements.txt
django==1.10.6
django-environ==0.4.1
psycopg2==2.7
requests==2.13.0
pyjwt==1.4.2
colorlog==2.10.0
djangorestframework==3.5.4
whitenoise==3.3.0
ipython==5.3.0
drfdocs==0.0.11
django-redis==4.7.0
arrow==0.10.0
